modificacion bigIncrements

+++ database/migrations/2019_08_02_184711_create_products_table.php
@@ -14,14 +14,14 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('name');
             $table->string('description');
             $table->text('long_description')->nullable();
             $table->float('price'); 
 
             //FK
-            $table->bigIncrements('category_id')->unsigned()->nullable();
+            $table->integer('category_id')->unsigned()->nullable();
             $table->foreign('category_id')->references('id')->on('categories');
 
             $table->timestamps();
